refactor: add comments, change structure
