+++ src/validation/ErrorHandler.java
@@ -2,7 +2,6 @@ package validation;
 
 public interface ErrorHandler<T> {
     void validate();
-
     void add(T field);
     void delete(T field);
 }
